+++ internal/transport/http/userHandler/dto/create.user.go
@@ -7,6 +7,6 @@ type CreateUserRequest struct {
 }
 
 type CreateUserResponse struct {
-	Id      int64  `json:"id" example:"42"`
+	Id      int    `json:"id" example:"42"`
 	Message string `json:"message" example:"user created"`
 }
